create logger API
